topology: disable force-gossip-topology-changes option
The patch marks force-gossip-topology-changes as deprecated and removes
tests that use it. There are two tests (test_different_group0_ids and
test_raft_fix_broken_snapshot) which are marked as xfail instead since
it looks like gossiper mode was used there as a way to easily achieve
a certain state, so more investigation is needed if the tests can be
fixed to use raft mode instead.
