+++ service/storage_service.cc
@@ -3189,9 +3189,6 @@ future<> storage_service::join_cluster(sharded<service::storage_proxy>& proxy,
                 throw std::runtime_error(
                         "Cannot start in the Raft-based recovery procedure - Raft-based topology has not been enabled");
             }
-            if (_db.local().get_config().force_gossip_topology_changes()) {
-                throw std::runtime_error("Cannot force gossip topology changes in the Raft-based recovery procedure");
-            }
         }
     }
 
@@ -3215,9 +3212,6 @@ future<> storage_service::join_cluster(sharded<service::storage_proxy>& proxy,
     } else if (_group0->joined_group0()) {
         // We are a part of group 0.
         set_topology_change_kind(upgrade_state_to_topology_op_kind(_topology_state_machine._topology.upgrade_state));
-        if (_db.local().get_config().force_gossip_topology_changes() && raft_topology_change_enabled()) {
-            throw std::runtime_error("Cannot force gossip topology changes - the cluster is using raft-based topology");
-        }
         slogger.info("The node is already in group 0 and will restart in {} mode", raft_topology_change_enabled() ? "raft" : "legacy");
     } else if (_sys_ks.local().bootstrap_complete()) {
         if (co_await _sys_ks.local().load_topology_features_state()) {
@@ -3238,13 +3232,8 @@ future<> storage_service::join_cluster(sharded<service::storage_proxy>& proxy,
 
         if (_group0->load_my_id() == g0_info.id) {
             // We're creating the group 0.
-            if (_db.local().get_config().force_gossip_topology_changes()) {
-                slogger.info("We are creating the group 0. Start in legacy topology operations mode by force");
-                set_topology_change_kind(topology_change_kind::legacy);
-            } else {
-                slogger.info("We are creating the group 0. Start in raft topology operations mode");
-                set_topology_change_kind(topology_change_kind::raft);
-            }
+            slogger.info("We are creating the group 0. Start in raft topology operations mode");
+            set_topology_change_kind(topology_change_kind::raft);
         } else {
             // Ask the current member of the raft group about which mode to use
             auto params = join_node_query_params {};
@@ -3252,9 +3241,6 @@ future<> storage_service::join_cluster(sharded<service::storage_proxy>& proxy,
                     &_messaging.local(), netw::msg_addr(g0_info.ip_addr), g0_info.id, std::move(params));
             switch (result.topo_mode) {
             case join_node_query_result::topology_mode::raft:
-                if (_db.local().get_config().force_gossip_topology_changes()) {
-                    throw std::runtime_error("Cannot force gossip topology changes - joining the cluster that is using raft-based topology");
-                }
                 slogger.info("Will join existing cluster in raft topology operations mode");
                 set_topology_change_kind(topology_change_kind::raft);
                 break;
