+++ test/cluster/suite.yaml
@@ -10,7 +10,6 @@ extra_scylla_config_options:
     tablets_mode_for_new_keyspaces: enabled
 run_first:
   - test_raft_recovery_stuck
-  - test_raft_recovery_basic
   - test_group0_schema_versioning
   - test_tablets_migration
   - test_zero_token_nodes_topology_ops
@@ -39,7 +38,6 @@ run_in_dev:
   - test_raft_ignore_nodes
   - test_group0_schema_versioning
   - test_different_group0_ids
-  - test_replace_ignore_nodes
   - test_zero_token_nodes_no_replication
   - test_not_enough_token_owners
   - test_replace_alive_node
