+++ test/cluster/test_table_drop.py
@@ -9,7 +9,6 @@ async def test_drop_table_during_streaming_receiver_side(manager: ManagerClient)
         'error_injections_at_startup': ['stream_mutation_fragments_table_dropped'],
         'enable_repair_based_node_ops': False,
         'enable_user_defined_functions': False,
-        'force_gossip_topology_changes': True,
         'tablets_mode_for_new_keyspaces': 'disabled'
     }) for _ in range(2)]
