+++ test/cluster/test_tablets.py
@@ -833,27 +833,6 @@ async def test_keyspace_creation_cql_vs_config_sanity(manager: ManagerClient, wi
         assert res is None
 
 @pytest.mark.asyncio
-async def test_tablets_and_gossip_topology_changes_are_incompatible(manager: ManagerClient):
-    cfg = {"tablets_mode_for_new_keyspaces": "enabled", "force_gossip_topology_changes": True}
-    with pytest.raises(Exception, match="Failed to add server"):
-        await manager.server_add(config=cfg)
-
-@pytest.mark.asyncio
-async def test_tablets_disabled_with_gossip_topology_changes(manager: ManagerClient):
-    cfg = {"tablets_mode_for_new_keyspaces": "disabled", "force_gossip_topology_changes": True}
-    await manager.server_add(config=cfg)
-    cql = manager.get_cql()
-    async with new_test_keyspace(manager, "WITH replication = {'class': 'NetworkTopologyStrategy', 'replication_factor': 1}") as ks_name:
-        res = cql.execute(f"SELECT * FROM system_schema.scylla_keyspaces WHERE keyspace_name = '{ks_name}'").one()
-        logger.info(res)
-
-    for enabled in ["false", "true"]:
-        expected = r"Error from server: code=2000 \[Syntax error in CQL query\] message=\"line 1:126 no viable alternative at input 'tablets'\""
-        with pytest.raises(SyntaxException, match=expected):
-            ks_name = unique_name()
-            await cql.run_async(f"CREATE KEYSPACE {ks_name} WITH replication = {{'class': 'NetworkTopologyStrategy', 'replication_factor': 1}} AND tablets {{'enabled': {enabled}}};")
-
-@pytest.mark.asyncio
 @pytest.mark.skip_mode(mode='release', reason='error injections are not supported in release mode')
 async def test_tablet_streaming_with_unbuilt_view(manager: ManagerClient):
     """
