+++ test/cluster/test_view_build_status.py
@@ -201,165 +201,6 @@ async def test_view_build_status_snapshot(manager: ManagerClient):
     await wait_for_view(cql, "vt1", 4)
     await wait_for_view(cql, "vt2", 4)
 
-# Start cluster in view_builder v1 mode and migrate to v2.
-# Verify the migration copies the v1 data, and the new v2 table
-# is used after the migration.
-@pytest.mark.asyncio
-async def test_view_build_status_migration_to_v2(request, manager: ManagerClient):
-    # First, force the first node to start in legacy mode
-    cfg = {'force_gossip_topology_changes': True, 'tablets_mode_for_new_keyspaces': 'disabled'}
-
-    servers = [await manager.server_add(config=cfg)]
-    # Enable raft-based node operations for subsequent nodes - they should fall back to
-    # using gossiper-based node operations
-    del cfg['force_gossip_topology_changes']
-
-    servers += [await manager.server_add(config=cfg) for _ in range(2)]
-
-    logging.info("Waiting until driver connects to every server")
-    cql, hosts = await manager.get_ready_cql(servers)
-
-    logging.info("Checking the upgrade state on all nodes")
-    for host in hosts:
-        status = await manager.api.raft_topology_upgrade_status(host.address)
-        assert status == "not_upgraded"
-
-    ks = await create_keyspace(cql)
-    await create_table(cql, ks)
-    await create_mv(cql, ks, "vt1")
-
-    # Verify we're using v1 now
-    v = await get_view_builder_version(cql)
-    assert v == 1
-
-    await wait_for_row_count(cql, "system_distributed.view_build_status", 3, hosts[0])
-
-    result = await cql.run_async("SELECT * FROM system.view_build_status_v2")
-    assert len(result) == 0
-
-    logging.info("Triggering upgrade to raft topology")
-    await manager.api.upgrade_to_raft_topology(hosts[0].address)
-
-    logging.info("Waiting until upgrade finishes")
-    await asyncio.gather(*(wait_until_topology_upgrade_finishes(manager, h.address, time.time() + 60) for h in hosts))
-
-    logging.info("Checking migrated data in system")
-
-    await asyncio.gather(*(wait_for(lambda: view_builder_is_v2(cql, host=h), time.time() + 60) for h in hosts))
-
-    # Check that new writes are written to the v2 table
-    await create_mv(cql, ks, "vt2")
-    await asyncio.gather(*(wait_for_view_v2(cql, ks, "vt2", 3, host=h) for h in hosts))
-
-    await wait_for_row_count(cql, "system.view_build_status_v2", 6, hosts[0])
-
-# Migrate the view_build_status table to v2 and write to the table during the migration.
-# The migration process goes through an intermediate stage where it writes to
-# both the old and new table, so the write should not be lost.
-@pytest.mark.asyncio
-async def test_view_build_status_migration_to_v2_with_write_during_migration(request, manager: ManagerClient):
-    # First, force the first node to start in legacy mode
-    cfg = {'force_gossip_topology_changes': True, 'tablets_mode_for_new_keyspaces': 'disabled'}
-
-    servers = [await manager.server_add(config=cfg)]
-    # Enable raft-based node operations for subsequent nodes - they should fall back to
-    # using gossiper-based node operations
-    del cfg['force_gossip_topology_changes']
-
-    servers += [await manager.server_add(config=cfg) for _ in range(2)]
-
-    logging.info("Waiting until driver connects to every server")
-    cql, hosts = await manager.get_ready_cql(servers)
-
-    logging.info("Checking the upgrade state on all nodes")
-    for host in hosts:
-        status = await manager.api.raft_topology_upgrade_status(host.address)
-        assert status == "not_upgraded"
-
-    ks = await create_keyspace(cql)
-    await create_table(cql, ks)
-
-    inj_insert = "view_builder_pause_add_new_view"
-    await manager.api.enable_injection(servers[1].ip_addr, inj_insert, one_shot=True)
-
-    await create_mv(cql, ks, "vt1")
-
-    # pause the migration between reading the old table and writing to the new table, so we have
-    # a time window where new writes may be lost.
-    # we don't know who the coordinator is so inject in all nodes.
-    inj_upgrade = "view_builder_pause_in_migrate_v2"
-    for s in servers:
-        await manager.api.enable_injection(s.ip_addr, inj_upgrade, one_shot=True)
-
-    logging.info("Triggering upgrade to raft topology")
-    await manager.api.upgrade_to_raft_topology(hosts[0].address)
-
-    logging.info("Waiting until upgrade finishes")
-    await asyncio.gather(*(wait_until_topology_upgrade_finishes(manager, h.address, time.time() + 60) for h in hosts))
-
-    logging.info("Checking migrated data in system")
-
-    # Now that the upgrade is paused, write the new view.
-    await manager.api.message_injection(servers[1].ip_addr, inj_insert)
-    await asyncio.sleep(1)
-
-    # continue the migration
-    for s in servers:
-        await manager.api.message_injection(s.ip_addr, inj_upgrade)
-
-    await asyncio.gather(*(wait_for(lambda: view_builder_is_v2(cql, host=h), time.time() + 60) for h in hosts))
-
-    await asyncio.gather(*(wait_for_view_v2(cql, ks, 'vt1', 3, host=h) for h in hosts))
-
-# Migrate the view_build_status table to v2 while there is an 'old' write operation in progress.
-# The migration should wait for the old operations to complete before continuing, otherwise
-# these writes may be lost.
-@pytest.mark.asyncio
-async def test_view_build_status_migration_to_v2_barrier(request, manager: ManagerClient):
-    # First, force the first node to start in legacy mode
-    cfg = {'force_gossip_topology_changes': True, 'tablets_mode_for_new_keyspaces': 'disabled'}
-
-    servers = [await manager.server_add(config=cfg)]
-    # Enable raft-based node operations for subsequent nodes - they should fall back to
-    # using gossiper-based node operations
-    del cfg['force_gossip_topology_changes']
-
-    servers += [await manager.server_add(config=cfg) for _ in range(2)]
-
-    logging.info("Waiting until driver connects to every server")
-    cql, hosts = await manager.get_ready_cql(servers)
-
-    logging.info("Checking the upgrade state on all nodes")
-    for host in hosts:
-        status = await manager.api.raft_topology_upgrade_status(host.address)
-        assert status == "not_upgraded"
-
-    ks = await create_keyspace(cql)
-    await create_table(cql, ks)
-
-    # Create MV and delay the write operation to the old table
-    inj_insert = "view_builder_pause_add_new_view"
-    await manager.api.enable_injection(servers[1].ip_addr, inj_insert, one_shot=True)
-    await create_mv(cql, ks, "vt1")
-
-    # The upgrade should perform a barrier and wait for the delayed operation to complete before continuing.
-    logging.info("Triggering upgrade to raft topology")
-    await manager.api.upgrade_to_raft_topology(hosts[0].address)
-
-    logging.info("Waiting until upgrade finishes")
-    await asyncio.gather(*(wait_until_topology_upgrade_finishes(manager, h.address, time.time() + 60) for h in hosts))
-
-    # the upgrade should now be waiting for the insert to complete.
-    # unpause the insert
-    await asyncio.sleep(1)
-    await manager.api.message_injection(servers[1].ip_addr, inj_insert)
-
-    logging.info("Checking migrated data in system")
-
-    await asyncio.gather(*(wait_for(lambda: view_builder_is_v2(cql, host=h), time.time() + 60) for h in hosts))
-
-    await asyncio.gather(*(wait_for_view_v2(cql, ks, 'vt1', 3, host=h) for h in hosts))
-
 # Test that when removing a node from the cluster, we clean its rows from
 # the view build status table.
 @pytest.mark.asyncio
@@ -427,74 +268,6 @@ async def test_view_build_status_with_replace_node(manager: ManagerClient):
         return True
 
     await wait_for(node_rows_replaced, time.time() + 60)
-
-# Start with view_build_status v1 mode, and create entries such that
-# some of them correspond to removed nodes or non-existent views.
-# Then migrate to v2 table and verify that only valid entries belonging to known nodes
-# and views are migrated to the new table.
-@pytest.mark.asyncio
-async def test_view_build_status_migration_to_v2_with_cleanup(request, manager: ManagerClient):
-    # First, force the first node to start in legacy mode
-    cfg = {'force_gossip_topology_changes': True, 'tablets_mode_for_new_keyspaces': 'disabled'}
-
-    servers = [await manager.server_add(config=cfg)]
-    # Enable raft-based node operations for subsequent nodes - they should fall back to
-    # using gossiper-based node operations
-    del cfg['force_gossip_topology_changes']
-
-    # We start with total 4 nodes and we will remove one of them before the migration.
-    servers += [await manager.server_add(config=cfg) for _ in range(3)]
-
-    logging.info("Waiting until driver connects to every server")
-    cql, hosts = await manager.get_ready_cql(servers)
-
-    logging.info("Checking the upgrade state on all nodes")
-    for host in hosts:
-        status = await manager.api.raft_topology_upgrade_status(host.address)
-        assert status == "not_upgraded"
-
-    # Create a view. This will insert 4 entries to the view build status table, one for each node.
-    ks_name = await create_keyspace(cql)
-    await create_table(cql, ks_name)
-    await create_mv(cql, ks_name, "vt1")
-
-    await wait_for_view_v1(cql, "vt1", 4)
-
-    await wait_for_row_count(cql, "system_distributed.view_build_status", 4, hosts[0])
-
-    # Insert a row that doesn't correspond to an existing view, but does correspond to a known host.
-    # This row should get cleaned during migration.
-    s0_host_id = await manager.get_host_id(servers[0].server_id)
-    await cql.run_async(f"INSERT INTO system_distributed.view_build_status(keyspace_name, view_name, host_id, status) \
-                          VALUES ('{ks_name}', 'view_doesnt_exist', {s0_host_id}, 'SUCCESS')")
-
-    # Remove the last node. the entry for this node in the view build status remains and it
-    # corresponds now to an unknown node. The migration should remove it.
-    logging.info("Removing last node")
-    await manager.server_stop_gracefully(servers[-1].server_id)
-    await manager.remove_node(servers[0].server_id, servers[-1].server_id)
-
-    servers = servers[:-1]
-    hosts = await wait_for_cql_and_get_hosts(cql, servers, time.time() + 60)
-
-    logging.info("Triggering upgrade to raft topology")
-    await manager.api.upgrade_to_raft_topology(hosts[0].address)
-
-    logging.info("Waiting until upgrade finishes")
-    await asyncio.gather(*(wait_until_topology_upgrade_finishes(manager, h.address, time.time() + 60) for h in hosts))
-
-    logging.info("Checking migrated data in system")
-
-    # Wait for migration and upgrade to view build status v2.
-    await asyncio.gather(*(wait_for(lambda: view_builder_is_v2(cql, host=h), time.time() + 60) for h in hosts))
-    await wait_for_view_v2(cql, ks_name, "vt1", 3)
-
-    # Verify that after migration we kept only the entries for the known nodes and views.
-    async def rows_migrated():
-        result = await cql.run_async("SELECT * FROM system.view_build_status_v2", host=hosts[0])
-        return (len(result) == 3) or None
-
-    await wait_for(rows_migrated, time.time() + 60)
 
 # Reproduces scylladb/scylladb#20754
 # View build status migration is doing read with CL=ALL, so it requires all nodes to be up.
